requirements.txt
numpy>=1.20
click>=8.0
pillow==8.3.1
scipy==1.7.1
requests==2.26.0
tqdm==4.62.2
ninja==1.10.2
matplotlib==3.4.2
imageio==2.9.0
imgui==1.3.0
glfw==2.2.0
pyopengl==3.1.5
imageio-ffmpeg==0.4.3
pyspng
psutil
mrcfile
tensorboard
einops
av